pyproject.toml
[project]
name = "de-ecommerce-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "apache-airflow==2.8.1",
    "dbt-core==1.9.0",
    "dbt-postgres==1.9.0",
    "faker>=37.12.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
